save allocation in OperatorSkipTimed
